chore: test tidying
