+++ src/app/shared/components/template/services/template-field.service.spec.ts
@@ -1,10 +1,6 @@
-import { HttpClientTestingModule } from "@angular/common/http/testing";
-import { TestBed } from "@angular/core/testing";
 import { TemplateFieldService } from "./template-field.service";
 import type { PromiseExtended } from "dexie";
 import { booleanStringToBoolean } from "src/app/shared/utils";
-import { ErrorHandlerService } from "src/app/shared/services/error-handler/error-handler.service";
-import { MockErrorHandlerService } from "src/app/shared/services/error-handler/error-handler.service.mock.spec";
 
 /** Mock calls for field values from the template field service to return test data */
 export class MockTemplateFieldService implements Partial<TemplateFieldService> {
@@ -26,19 +22,19 @@ export class MockTemplateFieldService implements Partial<TemplateFieldService> {
   }
 }
 
-describe("TemplateFieldService", () => {
-  let service: TemplateFieldService;
+// TODO - split from mock and add tests
 
-  beforeEach(async () => {
-    TestBed.configureTestingModule({
-      imports: [HttpClientTestingModule],
-      providers: [{ provide: ErrorHandlerService, useValue: new MockErrorHandlerService() }],
-    });
-    service = TestBed.inject(TemplateFieldService);
-    await service.ready();
-  });
-
-  it("should be created", () => {
-    expect(service).toBeTruthy();
-  });
-});
+// describe("TemplateFieldService", () => {
+// let service: TemplateFieldService;
+// beforeEach(async () => {
+//   TestBed.configureTestingModule({
+//     imports: [HttpClientTestingModule],
+//     providers: [{ provide: ErrorHandlerService, useValue: new MockErrorHandlerService() }],
+//   });
+//   service = TestBed.inject(TemplateFieldService);
+//   await service.ready();
+// });
+// it("should be created", () => {
+//   expect(service).toBeTruthy();
+// });
+// });
